Added Headless Mode, with example in /examples directory. Headless mode supports writing GIFs and a basic configuration via preprocessor definitions. Refer to headless example for usage.
Compiles jo_gif when compiling headless mode. See following link for source:
https://www.jonolick.com/uploads/7/9/2/1/7921194/jo_gif.cpp
This is used exclusively for GIF writing in headless mode, and is not
introduced in the standard configuration. The whole of the source is now
at the top of CTurtle.hpp, with include guards/etc.

The configuration allows for the same code to work in headless and
interactive mode, with very few differences between the two. The former
disallows any user input and background image loading.

Needs tested under MSVC.